refactor(memory): update CLI and chat to use memory-ops
