+++ src/chat-commands.ts
@@ -7,8 +7,8 @@ import { setConfigValue } from "./config";
 import type { ConfigScope } from "./config-contract";
 import { formatRelativeTime, nowIso } from "./datetime";
 import { t } from "./i18n";
-import { addMemory, listMemories, removeMemoryByPrefix } from "./memory";
 import type { MemoryScope } from "./memory-contract";
+import { addMemory, listMemories, removeMemoryByPrefix } from "./memory-ops";
 import { formatModel } from "./provider-config";
 import type { Session, SessionState, SessionTokenUsageEntry } from "./session-contract";
 import { findSkillByName } from "./skills";
